refactor(ansi): consolidate stripAnsi into shared util
Three divergent stripAnsi implementations existed (src/main/ansi-strip.ts,
a local copy in useSessionPreviews.ts, and another in model-detector.ts).
The renderer copy didn't strip OSC sequences, which leaked raw terminal
title-sequence text (e.g. "]0;my window title") into Grid tile previews.

- Move the canonical implementation to src/shared/ansi-strip.ts so it's
  importable from main, renderer, and shared code alike.
- Update history-manager.ts and classifier.ts (main) to import from the
  new shared path.
- Remove the local duplicate in useSessionPreviews.ts (renderer) and
  model-detector.ts (shared), both now importing the shared util —
  fixes the OSC-leak bug in Grid previews as a side effect.
- Add src/shared/ansi-strip.test.ts covering CSI color/SGR stripping,
  OSC sequences terminated by BEL and by ST, CSI private-mode sequences,
  \r\n and lone \r normalization, plain-text passthrough, and empty input.

Closes #69
